Fix incorrect move.transcode() method call
Transcoder options, called `ffmpeg_options` in the code, should be the 3rd argument. Options should be the 2nd argument, which can just be an empty hash, `{}`. Additionally, after correctly passing in the `ffmpeg_options` argument, `validate` needs to be set to `true`. When set to `false` the `transcoded_movie` variable will be returned as `nil` and the code will fail later on when that variable is referenced.